+++ Project/Game/GUI/Window/StatsWindow.cpp
@@ -114,7 +114,7 @@ void StatsWindow::drawGearIcons()
 		EquipmentComponent* equip = go->GetComponent<EquipmentComponent>();
 		if(equip)
 		{
-			for(int i = 0; i < 5; i++)
+			for(size_t i = 0; i < 5; i++)
 			{
 				switch(i)
 				{
